[BOPS-2278] Adds GitHub action to regenerate Dockerfiles
This adds a GitHub action to update `versions.json` will the latest Ruby
versions and regenerate the Dockerfiles using the `Dockerfile.template`.

It will run on a daily schedule and can be trigged manually from the
actions tab.

A separate action will be added later to build and push the images on a
less frequent schedule.