[BOPS-2306] Adds GitHub action to build, test, push docker images
This introduces a GitHub action that utilizes the auto-generated
Dockerfiles to build, test, and push ruby-jemalloc Docker images to GHCR
for our supported Ruby versions and Linux distributions.

The action is scheduled to run weekly on Monday mornings, following the
job that regenerates the Dockerfiles to account for new Ruby updates.

Note: GHCR was recommended by DevOps over DockerHub to reduce the number
of container registries we use. It also provides good support for
private images being accessed locally as well as in other GitHub Action
pipelines.