[BOPS-2737] Switch back to pushing to GHCR
This reverts several commits related to changing the registry from GHCR
(how this was initially designed) to Dockerhub, as we had some issues
at the time related to enabling public GitHub packages.

We are now focused on transitioning off of Dockerhub, so we are
returning to the original strategy.

Reverts:
- 6f2559a8c943f8bd30e4492051dd4baf483154ad
- 5cca3648735a9dd567709fcea14162146206a474